requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
Pillow
opencv-python
torch
torchvision
